requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.3
colorama==0.4.1
Django==2.2
django-cors-headers==2.5.3
django-filter==2.1.0
django-filters==0.2.1
djangorestframework==3.9.2
future==0.17.1
more-itertools==7.0.0
pluggy==0.9.0
py==1.8.0
pycparser==2.19
pytest==4.4.1
pytest-django==3.4.8
pytz==2018.9
serializers==0.2.4
six==1.12.0
sqlparse==0.3.0